o Now using explicit integers where ever possible, e.g. 1L instead of 1. o Make use of new Rdoc tag @usage everywhere.
git-svn-id: svn+ssh://svn.r-forge.r-project.org/svnroot/matrixstats/pkg/matrixStats@160 edb9625f-4e0d-4859-8d74-9fd3b1da38cb

+++ R/rowTabulates.R
@@ -12,8 +12,8 @@
 # }
 #
 # \usage{
-#   rowTabulates(x, values=NULL, ...)
-#   colTabulates(x, values=NULL, ...)
+#   @usage rowTabulates
+#   @usage colTabulates
 # }
 #
 # \arguments{
@@ -78,7 +78,7 @@ setMethod("rowTabulates", signature(x="matrix"), function(x, values=NULL, ...) {
 
 
   nbrOfValues <- length(values);
-  counts <- matrix(as.integer(0), nrow=nrow(x), ncol=nbrOfValues);
+  counts <- matrix(as.integer(0L), nrow=nrow(x), ncol=nbrOfValues);
   colnames(counts) <- names;
 
   dim <- dim(x);
